% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_data.R
\name{eo_data}
\alias{eo_data}
\title{imports data files from Eosense chambers}
\usage{
eo_data(path)
}
\arguments{
\item{path}{location of the data files}
}
\value{
a dataframe with continuous measurement with timestamps
}
\description{
imports data files from Eosense chambers
}
\examples{
path <- system.file("extdata/ex_data", package = "readosense")
eo_data(path)
}
