% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_import_log.R
\name{eo_import_log}
\alias{eo_import_log}
\title{imports log files from Eosense chambers}
\usage{
eo_import_log(
  path,
  closed = c(1:3),
  open = 0,
  fully_closed = 1,
  colnames = c("port", "valvestatus", "chamberstatus", "aux1", "aux2", "aux3", "aux4",
    "aux5", "temperaturev", "pressure"),
  nb_ports = 12
)
}
\arguments{
\item{path}{folder in which the log files are located}

\item{closed}{status numbers indicating chamber is closed}

\item{open}{status numbers indicating chamber is open}

\item{colnames}{column names as defined in Eosense documentation}

\item{nb_ports}{number of ports}
}
\description{
imports log files from Eosense chambers
}
