% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_log.R
\name{eo_log}
\alias{eo_log}
\title{imports log file from Eosense chambers}
\usage{
eo_log(
  path,
  colnames = c("port", "valvestatus", "chamberstatus", "aux1", "aux2", "aux3", "aux4",
    "aux5", "temperaturev", "pressure"),
  nb_ports = 12
)
}
\arguments{
\item{path}{folder in which the log files are located}

\item{colnames}{column names as defined in Eosense documentation}

\item{nb_ports}{number of ports}
}
\description{
imports log file from Eosense chambers
}
