% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_logs.R
\name{eo_logs}
\alias{eo_logs}
\title{extracts the measurements metadata from the logs}
\usage{
eo_logs(logs, time_buffer = 300)
}
\arguments{
\item{logs}{as provided by \link[readosense:eo_import_log]{eo_import_log}}

\item{buffer}{around the full closure of the chamber in seconds}
}
\value{
a dataframe with the measurement ids, port, and closing and opening
time
}
\description{
extracts the measurements metadata from the logs
}
