% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_env.R
\name{eo_env}
\alias{eo_env}
\title{extracts the environmental data from the logs}
\usage{
eo_env(logs)
}
\arguments{
\item{logs}{as provided by \link[readosense:eo_import_log]{eo_import_log}}
}
\value{
a dataframe with the datetime, air temperature (celsius), and
pressure (atm) for all chambers
}
\description{
extracts the environmental data from the logs
}
