% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_log.R
\name{eo_log}
\alias{eo_log}
\title{extracts the measurements metadata from the logs}
\usage{
eo_log(logs, time_buffer = 300)
}
\arguments{
\item{logs}{as provided by \link[readosense:eo_import_log]{eo_import_log}}

\item{buffer}{around the full closure of the chamber in seconds}
}
\description{
extracts the measurements metadata from the logs
}
